requirements.txt
flask==3.0.3
torch==2.3.1
torchvision==0.18.1
pillow==10.4.0
pickle-mixin==1.0.2
